# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Hub mentor API
  description: |
    A Flask + MongoDB API for the mentor service.
    
    ## Authentication
    All endpoints (except `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a JWT from the Developer Edition welcome / IdP flow, or use a
    test token from `test/e2e/e2e_auth.py` that matches your `JWT_SECRET` and expected claims.
  version: 0.1.0
  contact:
    name: Mentor Forge
    email: info@agile-learning-institute.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8391
    description: Local development server
...
paths:
...
  /api/Path:
    get:
      tags:
        - Path
      summary: Get all Path documents
      description: Retrieve infinite scroll batch of sorted, filtered Path documents
      operationId: getPaths
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Path
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Path
      summary: Create a new Path document
      description: Create a new Path document (requires staff or admin role)
      operationId: createPath
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathInput'
      responses:
        '201':
          description: Path created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Path/{PathId}:
    get:
      tags:
        - Path
      summary: Get Path by PathId
      description: Retrieve a specific Path document by its PathId
      operationId: getPath
      security:
        - bearerAuth: []
      parameters:
        - name: PathId
          in: path
          required: true
          description: The Path ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Path'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Path
      summary: Update Path document
      description: Update a Path document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updatePath
      security:
        - bearerAuth: []
      parameters:
        - name: PathId
          in: path
          required: true
          description: The Path ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathUpdate'
      responses:
        '200':
          description: Path updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Path'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Plan:
    get:
      tags:
        - Plan
      summary: Get all Plan documents
      description: Retrieve infinite scroll batch of sorted, filtered Plan documents
      operationId: getPlans
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Plan
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Plan
      summary: Create a new Plan document
      description: Create a new Plan document (requires staff or admin role)
      operationId: createPlan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanInput'
      responses:
        '201':
          description: Plan created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Plan/{PlanId}:
    get:
      tags:
        - Plan
      summary: Get Plan by PlanId
      description: Retrieve a specific Plan document by its PlanId
      operationId: getPlan
      security:
        - bearerAuth: []
      parameters:
        - name: PlanId
          in: path
          required: true
          description: The Plan ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Plan
      summary: Update Plan document
      description: Update a Plan document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updatePlan
      security:
        - bearerAuth: []
      parameters:
        - name: PlanId
          in: path
          required: true
          description: The Plan ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanUpdate'
      responses:
        '200':
          description: Plan updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by your identity provider (Developer Edition welcome flow for local dev).
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
...
    Path:
      type: object
      description: Path domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Path name
          pattern: '^[^\s]{1,40}$'
          example: my-Path
        description:
          type: string
          description: Path description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Path document
        status:
          type: string
          description: Path status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    PathInput:
      type: object
      description: Input for creating a Path document
      properties:
        name:
          type: string
          description: Path name
          pattern: '^[^\s]{1,40}$'
          example: my-Path
        description:
          type: string
          description: Path description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Path document
        status:
          type: string
          description: Path status
          enum:
            - active
            - archived
          example: active
    PathUpdate:
      type: object
      description: Fields to update in a Path document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Path name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Path description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Path status
          enum:
            - active
            - archived
    Plan:
      type: object
      description: Plan domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Plan name
          pattern: '^[^\s]{1,40}$'
          example: my-Plan
        description:
          type: string
          description: Plan description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Plan document
        status:
          type: string
          description: Plan status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    PlanInput:
      type: object
      description: Input for creating a Plan document
      properties:
        name:
          type: string
          description: Plan name
          pattern: '^[^\s]{1,40}$'
          example: my-Plan
        description:
          type: string
          description: Plan description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Plan document
        status:
          type: string
          description: Plan status
          enum:
            - active
            - archived
          example: active
    PlanUpdate:
      type: object
      description: Fields to update in a Plan document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Plan name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Plan description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Plan status
          enum:
            - active
            - archived
...